requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
scikit-learn>=1.2.0
faiss-cpu>=1.7.4
Pillow>=9.5.0
tqdm>=4.65.0
tensorboard>=2.12.0

# Vision and OCR
opencv-python>=4.7.0
pytesseract>=0.3.10
donut-python>=1.0.0

# Text processing
nltk>=3.8.1
spacy>=3.5.0
editdistance>=1.0.0

# Utilities
wandb>=0.15.0
jupyter>=1.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0 